backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.2
djangorestframework-simplejwt>=5.3
pymongo[srv]>=4.7
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
google-generativeai>=0.7
groq>=0.9
cloudinary>=1.41
django-cloudinary-storage>=0.3
Pillow>=10.0
python-dotenv>=1.0
